fix type

+++ workers/user-expiry.php
@@ -78,7 +78,7 @@ foreach ($SQL->getAllUserLastLogins() as $record) {
 $uid_to_idle_days = [];
 foreach ($uid_to_last_login as $uid => $last_login) {
     $seconds_since_last_login = $now - $uid_to_last_login[$uid];
-    $uid_to_idle_days[$uid] = round($seconds_since_last_login / (60 * 60 * 24));
+    $uid_to_idle_days[$uid] = (int) round($seconds_since_last_login / (60 * 60 * 24));
 }
 
 $pi_group_members = [];
